♻️ refactor: createFootprintGrid関数をresult-determinationに統合
- createFootprintGrid関数をresult-determination.ts内に移動
- footprint-grid.tsファイルを削除
- 関連する処理を同じファイル内にまとめて保守性を向上
- searchCrowPositions関数名を維持
- index.tsのエクスポートを適切に更新
