+++ src/logic/pathfinding/index.ts
@@ -5,5 +5,4 @@ export { findPath } from './main';
 export { bfsAllShortestPaths } from './bfs';
 
 // グリッド状態変更関数
-export { createFootprintGrid } from './footprint-grid';
 export { createRestTransitionGrid } from './rest-transition';
